+++ eRestaurant/Controllers/Master/RoleController.cs
@@ -27,6 +27,7 @@ namespace RocketPOS.Controllers.Master
 
         public ActionResult Index(int? noDelete)
         {
+            _iCommonService.GetPageWiseRoleRigths("Role");
             List<RoleModel> roleModel = new List<RoleModel>();
             roleModel = _iRoleService.GetRoleList().ToList();
             if (noDelete != null)
@@ -39,12 +40,19 @@ namespace RocketPOS.Controllers.Master
         public ActionResult Role(int? id)
         {
             RoleModel roleModel = new RoleModel();
-            if (id > 0)
+            if (UserRolePermissionForPage.Add == true || UserRolePermissionForPage.Edit == true)
             {
-                roleModel = _iRoleService.GetRoleById(Convert.ToInt32(id));
-            }
+                if (id > 0)
+                {
+                    roleModel = _iRoleService.GetRoleById(Convert.ToInt32(id));
+                }
 
-            return View(roleModel);
+                return View(roleModel);
+            }
+            else
+            {
+                return RedirectToAction("NotFound", "Error");
+            }
         }
 
         [HttpPost]
@@ -89,17 +97,23 @@ namespace RocketPOS.Controllers.Master
         public ActionResult Delete(int id)
         {
             int result = 0;
-            result = _iCommonService.GetValidateReference("WebRole", id.ToString());
-            if (result > 0)
+            if (UserRolePermissionForPage.Delete == true)
             {
-                return RedirectToAction(nameof(Index), new { noDelete = result });
+                result = _iCommonService.GetValidateReference("WebRole", id.ToString());
+                if (result > 0)
+                {
+                    return RedirectToAction(nameof(Index), new { noDelete = result });
+                }
+                else
+                {
+                    var deletedid = _iRoleService.DeleteRole(id);
+                    return RedirectToAction(nameof(Index));
+                }
             }
             else
             {
-                var deletedid = _iRoleService.DeleteRole(id);
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("NotFound", "Error");
             }
-            return RedirectToAction(nameof(Index));
         }
 
         private string ValidationRole(RoleModel roleModel)
